bunch of fixes for maximum upload size, output XSL transformation time in formtester report tab, fix bug with browsers that do not support filesystem API

+++ src/js/module/controller-webform.js
@@ -33,7 +33,7 @@ define( [ 'gui', 'connection', 'settings', 'enketo-js/Form', 'enketo-js/FormMode
             store = options.recordStore || null;
             fileManager = options.fileStore || null;
 
-            connection.init();
+            connection.init( true );
 
             if ( fileManager && fileManager.isSupported() && ( !store || store.getRecordList().length === 0 ) ) {
                 //clean up filesystem storage
@@ -444,7 +444,7 @@ define( [ 'gui', 'connection', 'settings', 'enketo-js/Form', 'enketo-js/FormMode
 
             //remove filesystem folder after successful submission
             $( document ).on( 'submissionsuccess', function( ev, recordName, instanceID ) {
-                if ( fileManager ) {
+                if ( fileManager && fileManager.isSupported() ) {
                     fileManager.deleteDir( instanceID );
                 }
                 if ( store ) {
